op-e2e: Extract out helper methods for connecting peers.
